fix(indexer): move fuction and fix imports
