+++ src/Database/Inscriptions/Utilities.ts
