+++ src/Libraries/Inscriptions/Utilities.ts
@@ -1,16 +1,7 @@
-import { parseLocation } from "../../Utilities/Transaction";
+import { getOutpointFromId } from "~Database/Inscriptions/Utilities";
 
-export function getIdFromOutpoint(outpoint: string) {
-  return outpoint.replace(":", "i");
-}
+import { parseLocation } from "../../Utilities/Transaction";
 
 export function getLocationFromId(id: string) {
   return parseLocation(getOutpointFromId(id));
-}
-
-export function getOutpointFromId(id: string) {
-  const outpoint = id.split("");
-  // Replace 65th character (i) with a colon (:)
-  outpoint[64] = ":";
-  return outpoint.join("");
 }
